✨ feat: Add support for object detection with DETR model
The DETR model is an encoder-decoder transformer with a convolutional backbone. Two heads are added on top of the decoder outputs in order to perform object detection.

---

feat: Add support for automatic speech recognition with S2T model

S2T is an end-to-end sequence-to-sequence transformer model. It is trained with standard autoregressive cross-entropy loss and generates the transcripts autoregressively.

---

feat: Add support for feature extraction with Vision Transformer (ViT) model

The Vision Transformer (ViT) is a transformer encoder model (BERT-like). Images are presented to the model as a sequence of fixed-size patches.

---

feat: Add support for image-to-text with GIT model

GIT is a Transformer decoder conditioned on both CLIP image tokens and text tokens.

---

feat: Add support for zero-shot image classification with CLIP models

Explore an alternative to ViT and ResNet (w/ AttentionPooling) CLIP models that scales well with model size and image resolution.